+++ bin/validate.js
@@ -2,15 +2,19 @@
 
 import { run, isCliError } from '../lib/cli.js';
 
-try {
-    run();
-} catch (e) {
-    // output user frendly message if cli error
-    if (isCliError(e)) {
-        console.error(e.message || e);
-        process.exit(2);
-    }
+async function perform() {
+    try {
+        await run();
+    } catch (e) {
+        // output user frendly message if cli error
+        if (isCliError(e)) {
+            console.error(e.message || e);
+            process.exit(1);
+        }
 
-    // otherwise re-throw exception
-    throw e;
-}
+        // otherwise re-throw exception
+        throw e;
+    }
+};
+
+perform();